% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keystats.R
\name{keystats}
\alias{keystats}
\title{keystats}
\usage{
keystats(key, n = 10000, top_num = 10, graph_col = "dodgerblue3")
}
\arguments{
\item{key}{keyword or hashtag that the user wishes to analyze}

\item{n}{number of tweets to analyze relating to key, maximum tweets to
analyze is 18,000, Default: 10000}

\item{top_num}{number of top words/hashtags/locations displayed in output,
Default: 10}

\item{graph_col}{color of output graphs, Default: 'dodgerblue3'}
}
\value{
an HTML output containing summary information and graphics of the
specified keyword(s) using input keyword(s). See details for specifics.
}
\description{
keystats() allows for easy Twitter analysis of user-specified
Twitter keywords and hashtags. See details for more information.
}
\details{
keystats() allows Twitter users to search the website by keyword or
hashtag,returning multiple graphics and statistical summaries such as a
bar plot of related unique words, a word cloud, a radar chart, and two
bar plots of location statistics.
}
\examples{
\dontrun{
if(interactive()){
 #keystats("data science", n=1000, top_num = 20)
 }
}
}
