% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userstats.R
\name{userstats}
\alias{userstats}
\title{userstats}
\usage{
userstats(
  user,
  n = 1000,
  bg_col = "white",
  txt_col = "#1A5276",
  year = 2019,
  x = 10
)
}
\arguments{
\item{user}{username of a real Twitter account that the user wishes to analyze}

\item{n}{number of tweets to analyze from that account, Default: 1000}

\item{bg_col}{background color of output tables, Default: 'white'}

\item{txt_col}{text color of output tables, Default: '#1A5276'}

\item{year}{year during which tweets are to be analyzed and displayed in
output, Default: 2019}

\item{x}{number of recent tweets to display from that account in the output,
Default: 10}
}
\value{
an HTML output containing summary information of the specified
account(s) using input username(s) as well as the specified number of recent
tweets from that account(s)
}
\description{
userstats() allows for easy Twitter analysis of user-specified
Twitter accounts. See details for more information.
}
\details{
userstats() allows Twitter users to search the website by username,
allowing the user to search multiple usernames at one time, specify which year
they would like summary information for, specify the number of tweets they
would like to analyze (with a maximum output allowed of 3,200 tweets),
customize the background color of output tables, and customize the text color
of the output with the assumption of knowledge of HTML color codes.
}
